move helper functions and vars
